Store only one of direct loads and transitive loads in `Package`.
`PackageSettings#precomputeTransitiveLoads` determines which is stored. If false (default), direct loads are stored as `ImmutableList<Module>`. If true, the full load graph is flattened and stored as `ImmutableList<Label>`. The only use case for precomputing transitive loads is when packages are expected to be serialized, since `Module` is too expensive to serialize. Otherwise, it is much more memory efficient to only store the direct loads and traverse the load graph on demand.

`Package#getStarlarkFileDependencies` is renamed to `getOrComputeTransitivelyLoadedStarlarkFiles`. Naively continuing to use this method in all cases results in a performance regression for certain queries using `buildfiles()` and/or `loadfiles()`. Instead, offer a new method `Package#visitLoadGraph` which permits callers to prune visitations across different packages. This actually results in a performance improvement vs prior state for these queries.

PiperOrigin-RevId: 532858099
Change-Id: Idd9acc135aa75dedfa8d2fac8be354a692eab313