Eliminate PackageArgument
The package() function is now implemented by a @StarlarkMethod rather than an anonymous subclass of StarlarkCallable.

The PackageArgument machinery is deleted. The convert-and-process approach is instead inlined into a series of if-elses for each argument (with the convert step using a helper to reduce repetition). The big if-else is factored out into an overridable method, so other implementations of package() can add their own parameters on top.

The package() symbol is now registered on the ConfiguredRuleClassProvider like any other BUILD toplevel. Removed its special casing in BazelStarlarkEnvironment, and the ability to register PackageArguments with the CRCP.

Fixed a copy-paste error where the "default_applicable_licenses" param would claim to be "default_package_metadata" for the purposes of error messages.

PiperOrigin-RevId: 532911239
Change-Id: I957354e0a4e32f7922b510c5537f56c6aff72ab9