Eliminate RuleClassProvider#getEnvironment
Callers who want to enumerate the built-in symbols of a .bzl file (not accounting for injection) can now call BazelStarlarkEnvironment#getUninjectedBuildBzlEnv instead. This previously wasn't possible in all contexts because BazelStarlarkEnvironment used to require a PackageFactory.

The new method also contains the (uninjected) `native` object. Updated some callers to account for this.

PiperOrigin-RevId: 533117799
Change-Id: I30d367356c123fad6c50faa0a9d712ce28102c98