Only track transitive packages when external repositories are enabled.
Storing transitive packages has a high memory cost. At the end of analysis, these are used to construct a mapping from package to source root. However, this can be accomplished by iterating over packages in the graph.

The remaining use cases for transitive packages are 1) bzlmod, which may need to construct a `RepoMappingManifestAction` with knowledge of transitive packages below a certain target (as opposed to the whole build) and 2) skymeld with external repository support, since it needs to plant a symlink forest as soon as each top level target is analyzed. I will collaborate with the owners of these modes to see whether this can be simplified.

A followup change will get rid of the `NestedSet<Package>` fields in the various `ConfiguredObjectValue` classes in favor of an alternate storage strategy, since those fields are now wasted when external repositories are disabled.

Using the value of `--enable_bzlmod` to determine whether to store transitive packages is safe even if that option is toggled, since a change to `StarlarkSemantics` invalidates packages.

PiperOrigin-RevId: 533154118
Change-Id: I40d8666290b624551c9daefb14db36ee745f0fa7