Move fixed Starlark env to StarlarkGlobals
I'm using the term "fixed environment" to refer to the Starlark symbols that are hardcoded to be present in the environment, as opposed to the ones that are there because they were registered on the rule class provider. For example, `select()` is in the fixed environment but `cc_common` is not.

This CL creates a new entry point, StarlarkGlobals, for determining the fixed environment. The idea is that BazelStarlarkEnvironment will defer to StarlarkGlobals to obtain the fixed environment, then add the symbols that were registered on the rule class provider, apply injection, etc. This replaces BazelStarlarkEnvironment's direct dependence on StarlarkLibrary. A follow-up CL will do the same for ConfiguredRuleClassProvider's dependence on StarlarkModules (and move that dependence to inside BazelStarlarkEnvironment).

StarlarkGlobals's implementation is in a separate package, similar to the split between RuleClassProvider and ConfiguredRuleClassProvider, to avoid adding new dependency edges from lib/analysis to lib/packages.

The actual implementation of StarlarkGlobals is tiny, and we can probably inline StarlarkModules and StarlarkLibrary into it in a follow-up.

Updated comments in BazelStarlarkEnvironment for pedantic accuracy.

PiperOrigin-RevId: 533213930
Change-Id: I9f5f3001bf6d3007b9372efb353ab1b46c0adcee