Make progress_reporting_test.sh more flexible.
Prior to this CL, we're asserting the counter to the "Still waiting for 1 job to
complete" message. The test is implicitly assuming that the workspace status
action would be the last one running (hence the `[3 / 4]` assertion).

This is fairly brittle since theoretically there's no guarantee that the
workspace status action would still be running after all the other actions in
the build has finished. This is especially true for Skymeld.

PiperOrigin-RevId: 533387894
Change-Id: Id100d7d96966a4f36b6031c19b389789938fc830