Make AndroidInstall actions hermetic with respect to dex shards.
They were missing from the inputs, but since the action had the LOCAL execution requirement it ran locally without sandboxing, so this bug could only be triggered by building without the bytes.

Also change the execution requirement to NO_REMOTE, which permits sandboxing and would have caught this issue even when building with the bytes.

Related to #14523.

PiperOrigin-RevId: 533549342
Change-Id: I731f15071e5e08822a298abc5146b91a56ba1071