Add CcSharedLibraryHintInfo to cc_common
It's not possible to check for the presence of top level symbols in Starlark
and take a code path or another depending on the presence since merely checking
for them will trigger an error.

For that reason, for users that want to migrate their code to use
CcSharedLibraryHintInfo and still support older Bazel versions it is necessary
that cc_common returns the provider so that the presence can be checked with
hasattr()

This will enable migrating upb_proto_library and not break builds with older
Bazel versions.

RELNOTES:none
PiperOrigin-RevId: 533636884
Change-Id: Icfb4deaada2f82ed8f7dd81274f35b8d23b50d79