Pass old max transitive source version to `SkyValueDirtinessChecker`s.
If `checkMtsvs` is true in `FilesystemValueChecker#getDirtyValues`, have `FilesystemValueChecker` retrieve the old `SkyValue`'s max transitive source version.

In `AbstractIncrementalInMemoryMemoizingEvaluator#inject`, only removes entries with no direct deps in `valuesToInject` whose values are equal to the present values in the graph AND have equal max transitive source versions. Since the mtsv's for the new and old `NodeEntry`s are always null, this is a no-op change for regular ol' incremental builds.

PiperOrigin-RevId: 534534109
Change-Id: Ibc1e5425b9fd0f7c52b2b888b56f4cd93ac3ce05