Inline `JavaInfo` internal providers as fields instead of a `TransitiveInfoProviderMap`
This makes progress towards rewriting `JavaInfo` in Starlark. The primary purpose of this (and the next few changes) is to first make the native provider resemble what the Starlark implementation will eventually look like. This has two benefits:
- Early identification of potential blockers - i.e. required refactoring or API changes
- Ensure we won't eventually cause a performance regression[^1]

This does make the class look worse but only temporarily - most of this code will go away once we switch over the Starlark provider.

PiperOrigin-RevId: 534754632
Change-Id: Ib5f776357b290e1f03e0c4e318c5a3c8bd331773