Promote the undeclared test outputs to an artifact.
The undeclared outputs can be produced in zipped or unzipped form, depending on the value of --zip_undeclared_test_outputs; test-setup.sh is responsible for the zipping. To simplify the code, the TestRunnerAction is declared as unconditionally producing an output directory, which will either contain a single zip file, a bunch of unzipped files, or be empty if no undeclared outputs were produced.

This change is required to make it possible to download the undeclared test outputs in unzipped form when building without the bytes.

Related to #17884.

PiperOrigin-RevId: 535188928
Change-Id: I94e72fe8840fa2274b66a2c8463cf90b4533c3a6