Add rule-transition to front of ConfiguredTargetFunction
Also memoizes computation of TargetAndConfiguration.

RELNOTES:
As a transitional step in a larger refactoring, rule transitions
are applied twice. Once during dependency resolution and once right before
analysis of those rules. After the refactoring is complete, rule transitions
will be applied only once.

Rule transitions are mostly expected to be idempotent, in which case this is
a no-op. For transitions where that is not the case, this could be a breaking
change or a cause a build performance regression.

PiperOrigin-RevId: 535843026
Change-Id: Icfad5a4a84081b7b63ca560ebed97c478cd6e2d4