In `RuleConfiguredTarget`, instead of storing an `ImmutableMap<Label, Artifact>` for outputs, find output artifacts using a linear search.
The only time we need to find an output `Artifact` by its `Label` is when creating an `OutputFileConfiguredTarget`. In such case, it is sufficient to just iterate over outputs to find one with a matching label. This is fast enough because most rules have a small number of actions, and most actions only have one output. In fact, `RuleConfiguredTarget#findArtifactByOutputLabel` does not even show up in a cpu profile even after being converted from a map lookup to a linear search.

The goal of this change was to cut the memory cost of the `ImmutableMap<Label, Artifact>`, but it also extends to a nice cleanup for setting generating action keys in `Actions.java`: a lot of the complexity was due to building the `ImmutableMap<Label, Artifact>`.

The `GeneratingActions` return type used to wrap the map with the list of actions is not needed - the list of actions is already passed in as a parameter and despite the method name, there was never any filtering done. Updated method names for better clarity.

PiperOrigin-RevId: 536488854
Change-Id: I6a06315cc0cb4a1c2d9b4449c386d63e949bdb7c