More aggressively reuse `CcToolchainVariables` in `CcToolchainProvider`.
Even with toolchain resolution enabled, the passed in configurations are often the same as the configuration used to construct the `CcToolchainProvider`. Reuse the provider's `CcToolchainVariables` any time the inputs to the construction match. This avoids the memory cost of creating a new instance on each call to `getBuildVariables()`.

For clarity, move the logic constructing `CcToolchainVariables` out of a helper class and into `CcToolchainProvider`. This makes it easier to see that the logic for reuse is correct.

PiperOrigin-RevId: 536795143
Change-Id: Ia2e3e665daac3e34fe897cd593fb4d8a623ac9b2