Make it able to inject non-remote file into RemoteInMemoryFileSystem.
In following CLs, we want to implement 'copy' semantic for `ctx.actions.symlink` in RemoteInMemoryFileSystem, i.e. instead of creating a symlink in `#createSymbolicLink`, we want to copy the target metadata to the link path.

There are cases where the target metadata is not a remote metadata, e.g. symlink to source, symlink to output of local action, etc. So we need to support inject non-remote metadata into RemoteInMemoryFileSystem.

PiperOrigin-RevId: 536994360
Change-Id: I1696db007304d638903b7250b8478491aef76795