Remove ProtoSourceInfo and use "_virtual_import" substring for import paths
This reduces heap memory usage, because we can use depsets of Files, instead of depsets of ProtoSourceInfo, which has a new instance of class with an Object array.

The effect of memory saving might be bigger in Bazel, because import_prefixes might be more commonly used.

The drawback is, that we can only strip prefixes using renaming to _virtual_imports subdirectory.

Alternatively we could keep ProtoSourceInfo and implement at least strip_import_prefix without renames. Measurements don't show any improvement in this case.

PiperOrigin-RevId: 537050475
Change-Id: Iff4ed7e0d4e4dfaab79449178c4b1278c1f369d0