Lockdown of j2objc_library by checking its new tag
j2objc_library will be deleted in the future and we would like to lock it down in order to stop spreading it.

Among relevant tests, j2objc_library_legacy will use 'old' j2objc_library since that's a temporary macro to help migration to the new j2objc_library.

Users will still be able to add this tag to their j2objc_library, but at least they will get notified that this library is locked and will be deleted in the future.

I've added this lockdown to the Native and the partially Starlarkified j2objc_library.

PiperOrigin-RevId: 537870824
Change-Id: I4a1bf05c3fb004b061353fc9925df972598adc4a