Export the original file and symbol of documented entities in starlark_doc_extract
It is a common pattern for modules to load a symbol (e.g. a rule or a provider)
from an internal implementation module and then re-export it inside a struct
namespace or even under a different name entirely. For building doc cross-
references, we want to be able to uniquely identify symbols, so in addition to
the names used inside a given module, we should export the moral equivalent of
the symbols' keys / fingerprints - the file where the symbol was originally
exported, and the name under which it was exported there.

Also take the opportunity to document advertised providers for rules.

As part of this change, we ought to have renamed ProviderNameGroup to something
like ProviderReferenceGroup; unfortunately, that would break existing Stardoc
markdown formatting templates, which require us to stay compatible with the
existing java proto API.

PiperOrigin-RevId: 537964225
Change-Id: Icb4c40448952edf7bbf93c5ddd616c26980dccf7