Warn on unsound directory inputs/outputs to any action, not just genrule actions.
The motivation behind this change is to eventually promote the warning to an error in the output case; this avoids surprising behavior such as described in https://github.com/bazelbuild/bazel/issues/18579. I'm less confident that the input case (source directories) can be similarly promoted to an error, but I'm still extending the warning to all actions for symmetry.

PiperOrigin-RevId: 538470433
Change-Id: Ic6c8a4cb353dda5f6397d286fe2e6faee28f896c